move ai dirs
