+++ server/ai/prompts.ts
